去除websocket睡眠时间

+++ v1/websocket/types.go
@@ -1,19 +1,19 @@
 package websocket
 
 type Data struct {
-	Ts        int     `json:"ts"`
+	Ts        int64   `json:"ts"`
 	Price     float64 `json:"price"`
 	Direction string  `json:"direction"`
 }
 
 type Tick struct {
 	Id   int    `json:"id"`
-	Ts   int    `json:"ts"`
+	Ts   int64  `json:"ts"`
 	Data []Data `json:"data"`
 }
 
 type TradeDetail struct {
-	Ts   int    `json:"ts"`
+	Ts   int64  `json:"ts"`
 	Ch   string `json:"ch"`
 	Tick Tick   `json:"tick"`
 }
